add linter and work flow for project
