+++ README.md
@@ -1,1 +1,49 @@
-# Node-Setup
+![](https://img.shields.io/badge/Microverse-blueviolet)
+
+# To-do list
+
+> In this project, I will build a simple yet functional To-do list. A To-do list is a tool that helps to organize your day. It simply lists the things that you need to do and allows you to mark them as complete. I will build a simple website that allows for doing that, and you I will do it using ES6 and Webpack!
+
+## Built With
+
+- JavaScript
+- HTML
+- CSS
+
+## Live Demo
+
+COMING SOON...
+
+## Getting Started
+
+To get a local copy up and running:
+
+Clone the repo
+git clone https://github.com/Itswali/toDoList.git
+Set it up localy.
+
+## Authors
+
+👤 **Wali Muhammad**
+
+- GitHub: [@Itswali](https://github.com/Itswali)
+- Twitter: [@Wali-Muhammad](https://twitter.com/WaliMuh94818599)
+- LinkedIn: [@wali-Muhammad](https://linkedin.com/in/wali-muhammad-666040244/)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](https://github.com/Itswali/toDoList/issues).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Not currently.
+
+## 📝 License
+
+This project is [MIT](./LICENSE) licensed.
