+++ src/index.html
@@ -9,7 +9,7 @@
     <h1>Hello webpack!</h1>
     <p>is webserver working</p>
     <h2>OH</h2>
-    <p>ss
+    <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Perspiciatis tempora, sed praesentium omnis aut dolorum, eaque molestiae error voluptatibus eos consectetur, corrupti quidem. Doloremque, et.
     </p>
 </body>
 </html>
