+++ src/style.css
@@ -1,0 +1,15 @@
+.hello {
+  color: blue;
+}
+
+body {
+  background-color: bisque;
+}
+
+h2 {
+  color: aqua;
+}
+
+p {
+  border: 2px solid black;
+}
