--- webpack-demo/src/style.css
@@ -1,15 +1,0 @@
-.hello {
-    color: blue;
-}
-
-body {
-    background-color: bisque;
-}
-
-h2 {
-    color: aqua;
-}
-
-p {
-    border: 2px solid black;
-}
